requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
dateparser==1.2.1
frozenlist==1.7.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
multidict==6.5.1
propcache==0.3.2
pycryptodome==3.23.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==20.0
pytz==2025.2
regex==2024.11.6
requests==2.32.4
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
